Extracted BuildScheduler and Waypointmanager from ZoopController
