+++ Zoop/Core/ZoopController.cs
@@ -1,7 +1,3 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Reflection;
 using Assets.Scripts.Inventory;
 using Assets.Scripts.Objects;
 using UnityEngine;
@@ -12,12 +8,11 @@ using ZoopMod.Zoop.Preview;
 namespace ZoopMod.Zoop.Core;
 
 /// <summary>
-/// Owns the live zoop session lifecycle, including start/cancel handling, waypoint management,
-/// preview refresh, and final build execution.
+/// Owns the live zoop session lifecycle, including start/cancel handling, preview refresh,
+/// and final build execution. Delegates build scheduling to <see cref="ZoopBuildScheduler"/>
+/// and waypoint mutation to <see cref="ZoopWaypointManager"/>.
 /// </summary>
-internal sealed class ZoopController(
-  ZoopPreviewValidator previewValidator,
-  ZoopPlacementUpdateGate placementUpdateGate)
+internal sealed class ZoopController : IZoopController
 {
   private enum ZoopLifecycleState
   {
@@ -26,33 +21,26 @@ internal sealed class ZoopController(
     PendingBuild
   }
 
-  private enum WaypointCaptureResult
-  {
-    Added,
-    Duplicate,
-    MissingCursor
-  }
-
-  private const float ConfirmDraftDelaySeconds = 0.100f;
-
-  private static readonly MethodInfo WaitUntilDoneMethod = typeof(InventoryManager).GetMethod("WaitUntilDone",
-    BindingFlags.NonPublic | BindingFlags.Instance, null,
-    [typeof(InventoryManager.DelegateEvent), typeof(float), typeof(Structure)],
-    null);
-
-  private readonly ZoopPreviewCoordinator previewCoordinator = new(previewValidator);
+  private readonly ZoopPlacementUpdateGate placementUpdateGate;
+  private readonly ZoopPreviewCoordinator previewCoordinator;
+  private readonly ZoopBuildScheduler buildScheduler;
+  private readonly ZoopWaypointManager waypointManager;
   private ZoopDraft activeDraft;
   private ZoopPreviewCache activePreviewCache;
   private ZoopLifecycleState state;
-  private Coroutine pendingBuildCoroutine;
-  private InventoryManager pendingBuildOwner;
-  private ZoopBuildPlan pendingBuildPlan;
-  private Coroutine buildExecutionCoroutine;
-  private InventoryManager buildExecutionOwner;
+
+  internal ZoopController(ZoopPreviewValidator previewValidator, ZoopPlacementUpdateGate placementUpdateGate)
+  {
+    this.placementUpdateGate = placementUpdateGate;
+    previewCoordinator = new ZoopPreviewCoordinator(previewValidator);
+
+    buildScheduler = new ZoopBuildScheduler(previewCoordinator, this);
+    waypointManager = new ZoopWaypointManager(previewCoordinator, this);
+  }
 
   public int PreviewCount => activeDraft?.PreviewCount ?? 0;
   public bool AllowPlacementUpdate => placementUpdateGate.AllowPlacementUpdate;
-  public bool IsBuildExecuting => buildExecutionCoroutine != null;
+  public bool IsBuildExecuting => buildScheduler.IsBuildExecuting;
   public bool IsZoopKeyPressed { get; set; }
   public bool IsZooping => state != ZoopLifecycleState.Idle;
   public bool IsPreviewing => state == ZoopLifecycleState.Previewing;
@@ -95,7 +83,7 @@ internal sealed class ZoopController(
     if (IsBuildExecuting)
     {
       ZoopLog.Debug("[Lifecycle] Canceling active zoop build execution.");
-      CancelBuildExecution();
+      buildScheduler.CancelBuildExecution();
       return;
     }
 
@@ -103,128 +91,19 @@ internal sealed class ZoopController(
     ResetSession(restoreCursorVisibility: true, cancelPendingBuild: true);
   }
 
-  public void ConfirmZoop(InventoryManager inventoryManager)
-  {
-    try
-    {
-      var confirmCoroutine = inventoryManager.StartCoroutine(ConfirmZoopAfterDelay(inventoryManager));
-      if (confirmCoroutine != null)
-      {
-        pendingBuildCoroutine = confirmCoroutine;
-        pendingBuildOwner = inventoryManager;
-        ZoopLog.Debug($"[Build] Waiting {ConfirmDraftDelaySeconds:0.###} seconds before confirming zoop draft.");
-        return;
-      }
-    }
-    catch (Exception exception)
-    {
-      ZoopLog.Error(exception, "[Build] Failed to schedule delayed zoop confirm.");
-      return;
-    }
+  public void ConfirmZoop(InventoryManager inventoryManager) => buildScheduler.ConfirmZoop(inventoryManager);
 
-    ZoopLog.Error("[Build] Unable to start delayed zoop confirm.");
-  }
+  public void AddWaypoint() => waypointManager.AddWaypoint();
 
-  private IEnumerator ConfirmZoopAfterDelay(InventoryManager inventoryManager)
-  {
-    yield return new WaitForSecondsRealtime(ConfirmDraftDelaySeconds);
+  public void RemoveLastWaypoint() => waypointManager.RemoveLastWaypoint();
 
-    ClearPendingBuildState();
-    FinalizeConfirmedZoop(inventoryManager);
-  }
+  public void CancelBuildExecution() => buildScheduler.CancelBuildExecution();
 
-  private void FinalizeConfirmedZoop(InventoryManager inventoryManager)
-  {
-    var draft = activeDraft;
-    if (!IsPreviewing || draft == null)
-    {
-      ZoopLog.Debug($"[Build] Confirm ignored. IsPreviewing={IsPreviewing}, DraftPresent={draft != null}.");
-      return;
-    }
-
-    if (!previewCoordinator.TryFinalizePreview(draft, activePreviewCache, inventoryManager))
-    {
-      ZoopLog.Debug("[Build] Confirm could not finalize the latest preview state; using the current draft preview.");
-    }
-
-    if (draft.HasError)
-    {
-      ZoopLog.Debug("[Build] Confirm ignored because the finalized zoop preview has errors.");
-      return;
-    }
-
-    var buildPlan = CaptureBuildPlan(draft, inventoryManager);
-    if (buildPlan.Count <= 0)
-    {
-      ZoopLog.Debug("[Build] Confirm produced an empty build plan; canceling zoop.");
-      CancelZoop();
-      return;
-    }
-
-    ZoopLog.Debug($"[Build] Confirmed zoop with {buildPlan.Count} planned piece(s).");
-    EnterPendingBuild(buildPlan);
-
-    if (!InventoryManager.IsAuthoringMode &&
-        InventoryManager.ConstructionCursor != null &&
-        InventoryManager.ConstructionCursor.BuildPlacementTime > 0.0)
-    {
-      if (TrySchedulePendingBuild(inventoryManager))
-      {
-        return;
-      }
-
-      ZoopLog.Error("[Build] Unable to start delayed zoop build; building immediately.");
-    }
-
-    BuildPendingZoop(inventoryManager);
-  }
-
-  public void AddWaypoint()
-  {
-    var draft = activeDraft;
-    var supportsWaypoints = ZoopConstructableRules.SupportsWaypoints(InventoryManager.ConstructionCursor);
-    if (!IsPreviewing || draft == null || !supportsWaypoints)
-    {
-      ZoopLog.Debug(
-        $"[Waypoint] Add ignored. IsPreviewing={IsPreviewing}, DraftPresent={draft != null}, SupportsWaypoints={supportsWaypoints}.");
-      return;
-    }
-
-    var currentPos = ZoopPreviewCoordinator.GetCurrentMouseGridPosition();
-    var lastWaypoint = draft.Waypoints[draft.Waypoints.Count - 1];
-    ZoopLog.Debug(
-      $"[Waypoint] Add requested. CurrentPos={(currentPos.HasValue ? currentPos.Value.ToString() : "<none>")}, LastWaypoint={lastWaypoint}, PreviewCount={draft.PreviewCount}, WaypointCount={draft.Waypoints.Count}.");
-    switch (TryCaptureCurrentWaypoint(draft, invalidatePreview: true, out var capturedPos))
-    {
-      case WaypointCaptureResult.Added:
-        ZoopLog.Debug(
-          $"[Waypoint] Added waypoint at {capturedPos} directly from the snapped cursor position. NewWaypointCount={draft.Waypoints.Count}.");
-        break;
-      case WaypointCaptureResult.Duplicate:
-        ZoopLog.Debug("[Waypoint] Add ignored because the current cursor position already matches the last waypoint.");
-        // TODO show message to user that waypoint is already added
-        break;
-      case WaypointCaptureResult.MissingCursor:
-        ZoopLog.Debug("[Waypoint] Add ignored because no current grid position was available.");
-        break;
-    }
-  }
-
-  public void RemoveLastWaypoint()
-  {
-    var draft = activeDraft;
-    if (!IsPreviewing || draft == null ||
-        !ZoopConstructableRules.SupportsWaypoints(InventoryManager.ConstructionCursor))
-    {
-      return;
-    }
-
-    if (draft.Waypoints.Count > 1)
-    {
-      draft.Waypoints.RemoveAt(draft.Waypoints.Count - 1);
-      previewCoordinator.Invalidate();
-    }
-  }
+  ZoopDraft IZoopController.ActiveDraft => activeDraft;
+  ZoopPreviewCache IZoopController.ActivePreviewCache => activePreviewCache;
+  void IZoopController.EnterPendingBuildState() => StopPreviewLoop(ZoopLifecycleState.PendingBuild);
+  void IZoopController.ResetSessionForBuild() => ResetSession(restoreCursorVisibility: false, cancelPendingBuild: false);
+  void IZoopController.ResumePreviewing(InventoryManager inventoryManager) => ResumePreviewing(inventoryManager);
 
   private void BeginZoop(InventoryManager inventoryManager, bool restartExisting = false)
   {
@@ -287,220 +166,6 @@ internal sealed class ZoopController(
     previewCoordinator.Start(draft, activePreviewCache, inventoryManager);
   }
 
-  private void BuildPendingZoop(InventoryManager inventoryManager)
-  {
-    var buildPlan = ConsumePendingBuildPlan();
-    if (buildPlan == null)
-    {
-      ZoopLog.Debug("[Build] Build callback ignored because no pending build plan remained.");
-      return;
-    }
-
-    ZoopLog.Debug($"[Build] Executing zoop build with {buildPlan.Count} piece(s).");
-    ResetSession(restoreCursorVisibility: false, cancelPendingBuild: false);
-
-    var buildCoroutine = inventoryManager.StartCoroutine(ExecuteBuildPlan(inventoryManager, buildPlan));
-    if (buildCoroutine == null)
-    {
-      ZoopLog.Error("[Build] Failed to start frame-sliced zoop build coroutine; canceling placement cursor.");
-      inventoryManager.CancelPlacement();
-      return;
-    }
-
-    buildExecutionOwner = inventoryManager;
-    buildExecutionCoroutine = buildCoroutine;
-  }
-
-  private static ZoopBuildPlan CaptureBuildPlan(ZoopDraft draft, InventoryManager inventoryManager)
-  {
-    var pieces = new List<ZoopBuildPiece>(draft.PreviewCount);
-    for (var structureIndex = 0; structureIndex < draft.PreviewCount; structureIndex++)
-    {
-      var previewPiece = draft.PreviewPieces[structureIndex];
-      var previewStructure = previewPiece.Structure;
-      var buildIndex =
-        ZoopConstructableResolver.ResolveBuildIndex(draft, inventoryManager, previewStructure, structureIndex);
-      if (buildIndex < 0)
-      {
-        ZoopLog.Error(
-          $"[Build] Unable to resolve build index for {previewStructure.PrefabName}; skipping zoop placement.");
-        continue;
-      }
-
-      var spawnPrefab = ResolveSpawnPrefabForBuild(draft, inventoryManager, buildIndex, previewStructure);
-      if (spawnPrefab == null)
-      {
-        ZoopLog.Error($"[Build] Unable to resolve spawn prefab for build index {buildIndex}; skipping zoop placement.");
-        continue;
-      }
-
-      pieces.Add(new ZoopBuildPiece(
-        spawnPrefab,
-        buildIndex,
-        previewStructure.transform.position,
-        previewStructure.transform.rotation));
-    }
-
-    return new ZoopBuildPlan(pieces);
-  }
-
-  private void CancelPendingBuild()
-  {
-    if (pendingBuildOwner != null && pendingBuildCoroutine != null)
-    {
-      ZoopLog.Debug("[Build] Stopping pending delayed zoop confirm/build coroutine.");
-      pendingBuildOwner.StopCoroutine(pendingBuildCoroutine);
-    }
-
-    ClearPendingBuildState();
-  }
-
-  public void CancelBuildExecution()
-  {
-    if (buildExecutionOwner != null && buildExecutionCoroutine != null)
-    {
-      ZoopLog.Debug("[Build] Stopping active zoop build coroutine.");
-      buildExecutionOwner.StopCoroutine(buildExecutionCoroutine);
-    }
-
-    ClearBuildExecutionState();
-  }
-
-  private void ClearPendingBuildState()
-  {
-    pendingBuildCoroutine = null;
-    pendingBuildOwner = null;
-    pendingBuildPlan = null;
-  }
-
-  private void ClearBuildExecutionState()
-  {
-    buildExecutionCoroutine = null;
-    buildExecutionOwner = null;
-  }
-
-  private ZoopBuildPlan ConsumePendingBuildPlan()
-  {
-    var buildPlan = pendingBuildPlan;
-    ClearPendingBuildState();
-    return buildPlan;
-  }
-
-  private void EnterPendingBuild(ZoopBuildPlan buildPlan)
-  {
-    StopPreviewLoop(ZoopLifecycleState.PendingBuild);
-    pendingBuildPlan = buildPlan;
-    ZoopLog.Debug($"[Lifecycle] Entered pending build with {buildPlan.Count} piece(s).");
-  }
-
-  private bool TrySchedulePendingBuild(InventoryManager inventoryManager)
-  {
-    if (WaitUntilDoneMethod == null || InventoryManager.ConstructionCursor == null)
-    {
-      return false;
-    }
-
-    try
-    {
-      var actionCoroutine = inventoryManager.StartCoroutine(WaitForPendingBuildCompletion(inventoryManager));
-
-      if (actionCoroutine == null)
-      {
-        return false;
-      }
-
-      pendingBuildCoroutine = actionCoroutine;
-      pendingBuildOwner = inventoryManager;
-      ZoopLog.Debug($"[Build] Scheduled native delayed build for {PreviewCount} preview piece(s).");
-      return true;
-    }
-    catch (Exception exception)
-    {
-      ZoopLog.Error(exception, "[Build] Failed to schedule delayed zoop build.");
-      return false;
-    }
-  }
-
-  private IEnumerator WaitForPendingBuildCompletion(InventoryManager inventoryManager)
-  {
-    if (state != ZoopLifecycleState.PendingBuild || pendingBuildPlan == null)
-    {
-      yield break;
-    }
-
-    yield return null;
-
-    if (state != ZoopLifecycleState.PendingBuild || pendingBuildPlan == null)
-    {
-      yield break;
-    }
-
-    if (InventoryManager.ConstructionCursor == null)
-    {
-      ZoopLog.Debug("[Build] Construction cursor vanished before native delayed build could start; resuming preview.");
-      var owner = pendingBuildOwner;
-      ClearPendingBuildState();
-      ResumePreviewing(owner);
-      yield break;
-    }
-
-    var onBuildFinished = new InventoryManager.DelegateEvent(() => BuildPendingZoop(inventoryManager));
-    var buildTime = InventoryManager.ConstructionCursor.BuildPlacementTime;
-    var structure = InventoryManager.ConstructionCursor;
-    var waitRoutine = WaitUntilDoneMethod.Invoke(
-      inventoryManager,
-      [onBuildFinished, buildTime, structure]) as IEnumerator;
-    if (waitRoutine == null)
-    {
-      ZoopLog.Error("[Build] Native delayed build coroutine was unavailable after start delay.");
-      var owner = pendingBuildOwner;
-      ClearPendingBuildState();
-      ResumePreviewing(owner);
-      yield break;
-    }
-
-    yield return waitRoutine;
-
-    // TODO Restore cursor here?
-
-    if (state != ZoopLifecycleState.PendingBuild || pendingBuildPlan == null)
-    {
-      yield break;
-    }
-
-    ZoopLog.Debug("[Build] Delayed zoop build was interrupted before completion; resuming zoop preview.");
-    var inventoryManagerAfterWait = pendingBuildOwner;
-    ClearPendingBuildState();
-    ResumePreviewing(inventoryManagerAfterWait);
-  }
-
-  private IEnumerator ExecuteBuildPlan(InventoryManager inventoryManager, ZoopBuildPlan buildPlan)
-  {
-    try
-    {
-      yield return ZoopBuildExecutor.BuildAll(inventoryManager, buildPlan);
-    }
-    finally
-    {
-      ClearBuildExecutionState();
-    }
-
-    ZoopLog.Debug($"[Build] Zoop build completed for {buildPlan.Count} piece(s); canceling placement cursor.");
-    inventoryManager.CancelPlacement();
-  }
-
-  private static Assets.Scripts.ICreativeSpawnable ResolveSpawnPrefabForBuild(ZoopDraft draft,
-    InventoryManager inventoryManager,
-    int buildIndex, Structure previewStructure)
-  {
-    if (InventoryManager.IsAuthoringMode && draft.ZoopSpawnPrefab != null)
-    {
-      return draft.ZoopSpawnPrefab;
-    }
-
-    return ZoopConstructableResolver.GetConstructableForBuildIndex(inventoryManager, buildIndex) ?? previewStructure;
-  }
-
   private void ResumePreviewing(InventoryManager inventoryManager)
   {
     if (inventoryManager == null ||
@@ -532,11 +197,11 @@ internal sealed class ZoopController(
       $"[Lifecycle] Resetting zoop session from state {state}. RestoreCursor={restoreCursorVisibility}, CancelPendingBuild={cancelPendingBuild}.");
     if (cancelPendingBuild)
     {
-      CancelPendingBuild();
+      buildScheduler.CancelPendingBuild();
     }
     else
     {
-      ClearPendingBuildState();
+      buildScheduler.ClearPendingBuildState();
     }
 
     StopPreviewLoop(ZoopLifecycleState.Idle);
@@ -560,32 +225,6 @@ internal sealed class ZoopController(
     {
       InventoryManager.ConstructionCursor.gameObject.SetActive(true);
     }
-  }
-
-  private WaypointCaptureResult TryCaptureCurrentWaypoint(ZoopDraft draft, bool invalidatePreview,
-    out Vector3 capturedPos)
-  {
-    capturedPos = default;
-
-    var currentPos = ZoopPreviewCoordinator.GetCurrentMouseGridPosition();
-    if (!currentPos.HasValue)
-    {
-      return WaypointCaptureResult.MissingCursor;
-    }
-
-    capturedPos = currentPos.Value;
-    if (ZoopPositionUtility.IsSameZoopPosition(draft.Waypoints[draft.Waypoints.Count - 1], capturedPos))
-    {
-      return WaypointCaptureResult.Duplicate;
-    }
-
-    draft.Waypoints.Add(capturedPos);
-    if (invalidatePreview)
-    {
-      previewCoordinator.Invalidate();
-    }
-
-    return WaypointCaptureResult.Added;
   }
 
   private static Structure GetSelectedConstructable(InventoryManager inventoryManager)
